Track neighbor counts when a cell is added
When a cell is added, add it as a live cell and add a dead cell with a
neighbor count of 1 for each neighbor location.  The idea is that
neighbors are calculated and kept up to date each time a cell is added.

At this point it will only work for properly for adding cells whose
neighbors don't conflict with any other live or dead cells.  That will
be the next step.